Move tests from g_sample to sample
